pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "src"
version = "0.0.1"
description = "Model to estimate vital signs"
authors = [
  { name = "Salil Apte" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    # Devtools
    "loguru",
    "mkdocs",
    "pip",
    "pytest",
    "python-dotenv",
    "ruff",
    "tqdm",
    "typer",
    "pre-commit",
    "pyarrow",
    "fastparquet",
    # Core ML / pipeline tools
    "numpy==1.25.0",
    "pandas==1.5.3",
    "scikit_learn==1.2.2",
    "xgboost",
    "polars",
    "scipy==1.10.1",
    "BaselineRemoval==0.1.4",
    "biosppy==1.0.0",
    # # # Core Deep Learning Libraries
    # "torch",  # PyTorch
    # "torchvision",  # PyTorch Vision
    # "torchaudio",  # PyTorch Audio
    # Hyperparameter tuning
    "optuna",
    # Visualization
    "plotly",
    "matplotlib==3.7.1",
    "seaborn==0.12.2",
    "pyppg",
    "joblib",
    "nbformat",
]
requires-python = "~=3.10.0"

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true


